+++ test/integration/CMakeLists.txt
@@ -12,11 +12,9 @@ create_standard_test(NAME terminator_integration SOURCES terminator.cpp)
 create_standard_test(NAME integrated_reaction_rates SOURCES integrated_reaction_rates.cpp)
 
 if(MICM_ENABLE_LLVM)
-  create_standard_test(NAME analytical_jit_rosenbrock_integration SOURCES analytical_jit_rosenbrock.cpp)
-  create_standard_test(NAME jit_terminator_integration SOURCES jit_terminator.cpp)
+  add_subdirectory(jit)
 endif()
 
 if(MICM_ENABLE_CUDA)
-  # the compute-sanitizer hangs for any integration test somehow; turns it off for now
-  create_standard_test(NAME analytical_cuda_rosenbrock_integration SOURCES analytical_cuda_rosenbrock.cpp LIBRARIES musica::micm_cuda)
+  add_subdirectory(cuda)
 endif()
