+++ test/integration/jit/CMakeLists.txt
@@ -6,6 +6,5 @@ include(test_util)
 ################################################################################
 # Tests
 
-create_standard_test(NAME jit_function SOURCES test_jit_function.cpp)
-add_subdirectory(process)
-add_subdirectory(solver)
+create_standard_test(NAME jit_analytical_rosenbrock_integration SOURCES test_jit_analytical_rosenbrock.cpp)
+create_standard_test(NAME jit_terminator_integration SOURCES test_jit_terminator.cpp)
