+++ test/integration/jit/test_jit_analytical_rosenbrock.cpp
@@ -1,8 +1,8 @@
-#include "analytical_policy.hpp"
-#include "analytical_surface_rxn_policy.hpp"
+#include "../analytical_policy.hpp"
+#include "../analytical_surface_rxn_policy.hpp"
 
-#include <micm/solver/jit_solver_builder.hpp>
-#include <micm/solver/jit_solver_parameters.hpp>
+#include <micm/jit/solver/jit_solver_builder.hpp>
+#include <micm/jit/solver/jit_solver_parameters.hpp>
 #include <micm/solver/rosenbrock_solver_parameters.hpp>
 
 #include <gtest/gtest.h>
