+++ test/regression/RosenbrockChapman/regression_test_jit_dforce_dy.cpp
@@ -1,8 +1,8 @@
 #include "jit_util.hpp"
 #include "regression_test_dforce_dy_policy.hpp"
 
-#include <micm/solver/jit_linear_solver.hpp>
-#include <micm/solver/jit_rosenbrock.hpp>
+#include <micm/jit/solver/jit_linear_solver.hpp>
+#include <micm/jit/solver/jit_rosenbrock.hpp>
 
 #include <gtest/gtest.h>
