+++ test/regression/RosenbrockChapman/regression_test_jit_p_force.cpp
@@ -1,8 +1,8 @@
 #include "jit_util.hpp"
 #include "regression_test_p_force_policy.hpp"
 
-#include <micm/solver/jit_linear_solver.hpp>
-#include <micm/solver/jit_rosenbrock.hpp>
+#include <micm/jit/solver/jit_linear_solver.hpp>
+#include <micm/jit/solver/jit_rosenbrock.hpp>
 #include <micm/util/vector_matrix.hpp>
 
 #include <gtest/gtest.h>
