+++ test/unit/CMakeLists.txt
@@ -1,6 +1,9 @@
 if(MICM_ENABLE_JSON)
   add_subdirectory(configure)
 endif()
+if(MICM_ENABLE_CUDA)
+  add_subdirectory(cuda)
+endif()
 if(MICM_ENABLE_LLVM)
   add_subdirectory(jit)
 endif()
