+++ test/unit/cuda/process/test_cuda_process_set.cpp
@@ -1,7 +1,7 @@
-#include <micm/process/cuda_process_set.hpp>
+#include <micm/cuda/process/cuda_process_set.hpp>
 #include <micm/process/process_set.hpp>
-#include <micm/util/cuda_dense_matrix.hpp>
-#include <micm/util/cuda_sparse_matrix.hpp>
+#include <micm/cuda/util/cuda_dense_matrix.hpp>
+#include <micm/cuda/util/cuda_sparse_matrix.hpp>
 #include <micm/util/matrix.hpp>
 #include <micm/util/sparse_matrix_standard_ordering.hpp>
 #include <micm/util/sparse_matrix_vector_ordering.hpp>
