+++ test/unit/cuda/solver/test_cuda_rosenbrock.cpp
@@ -1,11 +1,11 @@
 #include <micm/process/arrhenius_rate_constant.hpp>
-#include <micm/solver/cuda_rosenbrock.cuh>
-#include <micm/solver/cuda_rosenbrock.hpp>
-#include <micm/solver/cuda_solver_builder.hpp>
-#include <micm/solver/cuda_solver_parameters.hpp>
+#include <micm/cuda/solver/cuda_rosenbrock.cuh>
+#include <micm/cuda/solver/cuda_rosenbrock.hpp>
+#include <micm/cuda/solver/cuda_solver_builder.hpp>
+#include <micm/cuda/solver/cuda_solver_parameters.hpp>
 #include <micm/solver/rosenbrock.hpp>
-#include <micm/util/cuda_dense_matrix.hpp>
-#include <micm/util/cuda_sparse_matrix.hpp>
+#include <micm/cuda/util/cuda_dense_matrix.hpp>
+#include <micm/cuda/util/cuda_sparse_matrix.hpp>
 #include <micm/util/matrix.hpp>
 #include <micm/util/sparse_matrix.hpp>
 #include <micm/util/sparse_matrix_vector_ordering.hpp>
