+++ test/unit/cuda/util/test_cuda_sparse_matrix.cpp
@@ -1,8 +1,8 @@
 #include "cuda_matrix_utils.cuh"
-#include "test_sparse_matrix_policy.hpp"
+#include "../../util/test_sparse_matrix_policy.hpp"
 
-#include <micm/util/cuda_sparse_matrix.hpp>
-#include <micm/util/sparse_matrix_vector_ordering.hpp>
+#include <micm/cuda/util/cuda_sparse_matrix.hpp>
+#include <micm/cuda/util/sparse_matrix_vector_ordering.hpp>
 
 #include <gtest/gtest.h>
