+++ test/unit/jit/process/CMakeLists.txt
@@ -6,6 +6,5 @@ include(test_util)
 ################################################################################
 # Tests
 
-create_standard_test(NAME jit_function SOURCES test_jit_function.cpp)
-add_subdirectory(process)
-add_subdirectory(solver)
+create_standard_test(NAME jit_process_set SOURCES test_jit_process_set.cpp)
+create_standard_test(NAME jit_forcing_calculation SOURCES test_jit_forcing_calculation.cpp forcing_calculation.cpp)
