+++ test/unit/jit/solver/test_jit_linear_solver.cpp
@@ -1,6 +1,6 @@
-#include "test_linear_solver_policy.hpp"
+#include "../../solver/test_linear_solver_policy.hpp"
 
-#include <micm/solver/jit_linear_solver.hpp>
+#include <micm/jit/solver/jit_linear_solver.hpp>
 #include <micm/util/matrix.hpp>
 #include <micm/util/sparse_matrix.hpp>
 #include <micm/util/sparse_matrix_vector_ordering.hpp>
