+++ test/unit/process/CMakeLists.txt
@@ -15,13 +15,3 @@ create_standard_test(NAME ternary_chemical_activation_rate_constant SOURCES test
 create_standard_test(NAME troe_rate_constant SOURCES test_troe_rate_constant.cpp)
 create_standard_test(NAME tunneling_rate_constant SOURCES test_tunneling_rate_constant.cpp)
 create_standard_test(NAME user_defined_rate_constant SOURCES test_user_defined_rate_constant.cpp)
-
-# GPU tests
-if(MICM_ENABLE_CUDA)
-  create_standard_test(NAME cuda_process_set SOURCES test_cuda_process_set.cpp LIBRARIES musica::micm_cuda IS_CUDA_TEST)
-endif()
-
-if(MICM_ENABLE_LLVM)
-  create_standard_test(NAME jit_process_set SOURCES test_jit_process_set.cpp)
-  create_standard_test(NAME jit_forcing_calculation SOURCES test_jit_forcing_calculation.cpp forcing_calculation.cpp)
-endif()
